Boyscout: remove ActiveSupport::Concern since it's not *exactly* what we're trying to do
We only need the instance methods to be available in instances of
classes that specify the usage of Statisfaction, while
ActiveSupport::Concern proactively adds all methods to all classes that
include the module. However, since the including class is Object, we do
not want to 'pollute' this global namespace with our methods.